+++ CHANGES.md
@@ -1,7 +1,7 @@
 Poodinis Changelog
 ==================
 **Version NEXT**
-* ADD setting persistent registration options
+* ADD setting persistent registration and resolve options
 * DEPRECATE DO_NOT_ADD_CONCRETE_TYPE_REGISTRATION, use doNotAddConcreteTypeRegistration instead
 * DEPRECATE supplying register()'s registration options as variadic arguments. Use register(SuperType, ConcreteType)(RegistrationOption[]) instead.
 * ADD resolve options to container resolve()